Camera2: Fix JPEG quality metadata type, new warnings.: DO NOT MERGE
- JPEG quality tag is now a uint8_t, not an int. Update parameter code
accordingly.

- Fix new warnings about narrowing conversions.

Bug: 7944244

+++ services/camera/libcameraservice/camera2/Parameters.h
@@ -55,7 +55,7 @@ struct Parameters {
     int pictureWidth, pictureHeight;
 
     int32_t jpegThumbSize[2];
-    int32_t jpegQuality, jpegThumbQuality;
+    uint8_t jpegQuality, jpegThumbQuality;
     int32_t jpegRotation;
 
     bool gpsEnabled;
